Separated out the random part
